+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# 1.0
-Initial release
+# 1.1.0
+- Added support for character entities in format strings
+- Added `get` and `has` at-map format string functions
+
+# 1.0.0
+- Initial release